+++ OndselSolver/GESpMatParPvMarkoFast.cpp
@@ -57,7 +57,7 @@ void GESpMatParPvMarkoFast::doPivoting(size_t p)
 	SpRowDsptr spRowi;
 	rowPositionsOfNonZerosInPivotColumn->clear();
 	auto lookForFirstNonZeroInPivotCol = true;
-	i = m - 1;
+	i = (int)m - 1;
 	while (lookForFirstNonZeroInPivotCol) {
 		spRowi = matrixA->at(i);
 		if (spRowi->find(p) == spRowi->end()) {
